adapt: common: Introduce get_func_addr()
On some distributions with some games, "objdump -D" is not able to
resolve the "*@plt" symbols. So read the relocation address of the
required symbol with "objdump -R". Then, read the code address of
the jump to that address with "objdump -D" from the .plt section.

Example call:
get_func_addr /usr/games/warzone2100 _Znwm
echo "FUNC_ADDR: $FUNC_ADDR"

Noticed on Ubuntu 16.04 Xenial with warzone2100 3.1.1.