cfgparser: Add heap check support
E.g. in the warzone2100 config there are pointers on the heap which
we blindly follow. We need to know if they are at an address >= heap
start and < heap end. Those values differ a lot depending on ASLR,
PIE vs. non-PIE, etc. So allow a check value "e heap". Allow it only
as the first and only check in a check entry for now.

SFuncPtr 0x128 p StructFunct always
check 0x128 p e heap

Set a heap check flag and translate this to:

SFuncPtr 0x128 p StructFunct always
check 0x128 p e 0 g 0
check 0x128 p l 0

This is a check which will always fail. The values have to replaced
live with the heap start and the heap end. Also set an option flag
to show that we have heap checks in the current config.