ugtrain: PIE: Skip adding exe offset for values on stack
The function handle_statmem_pie() is called before gathering
__libc_stack_end and values on the stack are treated like static
memory in there. This leads to a wrong address for stack values
with a position-independent executable. So skip adding the exe
offset if the on_stack flag is set.

Found by code reading. This is untested as the game endless-sky
is not compiled as a PIE on openSUSE Leap 42.3 and on Ubuntu 16.04.

Fixes: 2ec26505b823 ("ugtrain: Add value on stack support")