ugtrain: Add heap check support
If there are heap checks, then read the memory regions of the
target process every cycle and determine the heap start and end
from them. Then write the heap limits to the values of the heap
checks. Do all that at the end of a cycle for not delaying time
critical dynamic memory handling.
Also rename handle_pie() to handle_aslr() as that function does
not only handle PIE but also early PIC.