examples: Move the dangerdeep config out of incomplete
It turned out that it is pretty feasible to adapt the dangerdeep
config. The object size 1784 bytes is quite rare. There are only
three occurrences and the first one is the one we are looking for.
Using it on openSUSE Leap 42.3 instead of Ubuntu 14.04 did not
require any modification but the _Znwm code address of the Submarine
object. So move the config out of incomplete.