cfgparser: checks: Take over address of current cfg_en as cfg_ref
Checks are allowed to use a config entry reference as their address.
The value of the config entry is compared to the check value. The
problem if using this with the current config entry is that the
check is processed before the current config entry gets its value.

So take over the address of the current config entry instead and
remove the cfg_ref if the found cfg_ref equals the current config
entry. Handle config entries on the stack as well.

Allows changing:
Credits stack 0x4a0 i64 l 500000000 A,X a
check stack 0x4a0 i64 g 1000

to:
Credits stack 0x4a0 i64 l 500000000 A,X a
check Credits i64 g 1000