cfgparser: checks: Use the reference address if 'addr' is read
Checks from config entry reference usually use the value of that
reference. An exception is if this points to the current config
entry. But the disadvantage is that the reference has to be always
enabled. Otherwise, it and all checks relying on it don't get their
value. Using the address of a config entry in checks as well solves
this but duplicates the address making it harder to adapt the config.

So introduce the check address keyword 'addr' in front of a reference
to indicate that the address of the reference has to be taken over.

Example:
Hull 0x9a0 f64 max 1,0 a
Shields 0x998 f64 max 1,0 a
check addr Hull f64 g 2

Translates to:
Hull 0x9a0 f64 max 1,0 a
Shields 0x998 f64 max 1,0 a
check 0x9a0 f64 g 2