cfgparser: checks: Introduce 'this' keyword
Using the keyword 'this' has the advantage that the config entry
name is not duplicated. Furthermore, it has advantages when it is
used in single-line macros with the 'define' keyword.

Also reduce code duplication by introducing a helper function which
takes over the address of a config entry.

Allows to change:
Credits stack 0x4a0 i64 l 500000000 A,X a
check Credits i64 g 1000

To:
Credits stack 0x4a0 i64 l 500000000 A,X a
check this i64 g 1000