libmemdisc: Drop explicit stream buffer locking
The bug in glibc seems to be fixed so that it is not required any
more to do explicit FILE pointer locking. Furthermore, with the
single flush at exit the potential race window is much smaller now.
So use the regular fflush() and fputs() functions now.

Reverts 91f418cf5531 ("libmemdisc: add locking to the output stream buffer")