ugtrain: Convert from getenv() to secure_getenv()
Coverity is still not happy with the getenv() handling. The GNU
specific function secure_getenv() exists long enough now. So use
that one and drop other code attempting to mitigate insecure
getenv() handling. Also move getting the home directory to the
system functions to be able to provide a Windows equivalent in
future.