cfgoutput: Reduce number of head lines
There are too many duplicate lines for the same dynmem object,
ptrmem entry and static memory region. So keep the pointer of the
previous config entry and check if this is still the same object or
memory region to be printed. Skip printing the head line in that
case. Also amend function names and variable names in the path for
printing the active config to make it clear that those are
processing the active config and not the general config.