configure: Add "--disable-procmem" option
Reading from /proc/$pid/mem has a huge performance advantage over
using ptrace() as a lot less syscalls are required. For cross
compiling we cannot detect at configure time if /proc/$pid/mem
will be available. So it is best to enable it by default and
providing a configure option to disable it. This is a lot easier
than what scanmem does.

Reference: #52