ugtrain: Cache dynmem target memory
Ptrace() handling is inefficient. There are quite some duplicate
and unnecessary calls depending on the number of checks and value
size/density. Fortunately, the number is still low enough so that
performance issues haven't been visible so far. With the Meltdown
mitigation introducing around 200ns additional delay for syscalls,
the number of syscalls in the hot path should really be optimized.

So introduce a cache list to dynmem entries and pointers to 8 byte
caches for all config/check entries. Allocate the first cache entry
when allocating the dynmem entry and move it to the lowest offset
in the dynmem block. Then copy address entries for the whole dynmem
config into an address list, sort it and build up further caches
from there. Assign the caches to the config/check entries at the end
and print them during config output for debugging.
In the hot path invalidate all caches when switching the current
object. Then just check if the cache is invalid before filling it
and read from the cache. Mark it as dirty if a value is changed and
write back all caches at the end.

This saves about 50..70% ptrace() calls in common dynmem configs.
It also makes the strace output more readable.

Reference: #52