ugtrain: Introduce /proc/$pid/mem I/O in 2K chunks
To reduce the number of syscalls for dynmem entries further, read
and write the caches in 2K chunks from/to /proc/$pid/mem the same
way as scanmem does. Fail incomplete reads/writes for now as those
are really unlikely.
Add the procmem file descriptor to the options, pass it over to
memattach_test(), and open the file. Pass the fd as the "pid"
parameter to the active config processing all the way through to
target memory reading and writing with pread() and pwrite(). Also
pass the cache size down.

Reference: #52