ugtrain: Fix the dump with procmem support
When passing the /proc/$pid/mem file descriptor down to the active
config processing, it has been forgotten to pass it down to the
dumping of dynamic and pointer memory objects. So do that now.
Also add the missing detaching if the fd cannot be opened in
memattach_test().

Reference: #52
Fixes: 24fccb8faf87 ("ugtrain: Introduce /proc/$pid/mem I/O in 2K chunks")