options: Convert to C++
The struct app_options is passed to many C++ functions. It was
originally intended to contain only command line options but
now it is also used to pass global variables down. The major
disadvantage of keeping the options as C code is that C++ classes
cannot be added to the options. So rename options.c to options.cpp,
change that struct to public class "Options" and fix all occurrences.
Also update ugtrain copyright information.