ugtrain: Cache static and stack target memory
Do the same for static memory which has been done for dynamic
memory already. Put the cache list into opt and pass opt down
additionally. Change the cache lists to pointers and allocate them
as needed so that the memset() for opt does not break things. Also
introduce new macros to avoid code duplication. Finally remove
variable "buf" from change_mem_val() as everything writes back to
a cache now.
Use the same cache list for static and stack memory values for now
as it is unlikely that those get mixed up (on x86_64 PIE .data
offsets start at 0x200000).
When handling PIE and the stack end, calculate the new cache offset
only once by doing it only if the dirty flag is not set, then set
the dirty flag. Clear the dirty flags for all caches in the list at
the end again. This ensures that caches are read from the correct
memory addresses.

Reference: #52