ugtrain: Split stack from statmem cache
Having both kinds of memory in the same cache list does not conflict
now but when implementing library/PIC support for static memory, this
will become a problem.
So introduce a new class StackOpt and put another cache list into it,
add a memory type parameter to macro WRITE_CACHES(), add a check if
the current cache list is used to OUTPUT_CACHE(), and handle the
stack cache list separately.

Reference: #52