ugtrain: Drop debian packaging and update version
Packaging should always be done separately from the upstream source.
The packaging for the (hacked) Nokia N9 is really obsolete now. So
drop that. Also update the version to 0.4~pre as there won't be a
version 0.3.1 as too many new features have been implemented since
finding the new development direction.
Also update the PACKAGE_BUGREPORT URL as nobody is using the mailing
list any way. The GitHub URL is also better for SEO.