aslr: Introduce helper function get_regions()
There is quite some code duplication to get the exe path and to
prepare reading the memory regions. So put that into a new inline
function get_regions(). Parameters are only the pid and the region
list. Also invert logic when finding memory regions to reduce
indentation and to improve readability this way.