ugtrain: Drop --glc option and update doc
The --glc option is a convenience option for easy video recording
of games running with ugtrain. But glc is unmaintained, the cursor
cannot be recorded and sound recording does not work. The better way
to record games is to use ffmpeg instead. So drop that --glc option
and update documentation mentioning it.