examples: btanks: Fix and enhance with parent pointer
Kicking out the weapon objects with no ammo was a bad idea as
secondary weapons are already allocated and do not work any more
this way. So revert this. Use the parent pointer instead and check
if the parent has a health value greater than possible without
cheating. This is hacky but works pretty well.

Revert "examples: btanks: Kick out non-player objects"
This reverts commit c19ab12051138a6a9ac54e6a181cd801bbfd8c28.