ugtrain: Enhance the cleanup with atexit()
For restoring the terminal, there is atexit() handling any way. So
instead of calling the cleanup function manually, it makes sense
to register it with atexit().
The problem is that the global variables of ugtrain must be available
to the atexit() handler. So introduce a new file "util.cpp" and
define those as static variables. Also introduce a getter function
to get those for the ugtrain main() function and to initialize them
in opt. Furthermore, fix the biggest one-time leaks reported by
Valgrind during discovery and split a clear_config() function off.
Fix a name conflict by renaming struct map to struct region_map and
of cause make the cleanup safer, so that it can be called early.