ugtrain: Clear config lines and use globals in opt
Move the config lines to util as a static variable and pass it back
to ugtrain main(). Clear the config lines vector when cleaning up
ugtrain and after running the adaptation. Rename it from "lines" to
"cfg_lines" and reduce the number of function parameters by making
use of the globals in opt. Introduce a new util function
init_opt_globals(), so that the options initialization can set them
again after zeroing all options.