discovery: Drop freeing rlist to fix the backtrace
The stage 3/4 backtrace is broken as the regions list is empty at
the post-processing step. So remove clearing the regions list in the
middle of the discovery.
The right way would be to add an async-signal-safe signal handler
which sets a flag to inform the forked off process running
run_stage1234_loop() that it should exit regularly.

Fixes: 2b6f291d1aee ("ugtrain: Enhance the cleanup with atexit()")