discovery: Exit the worker regularly
The discovery worker pumps data from the FIFO into the dynmem
discovery file. When killing it with SIGTERM from the parent, then
it exits from the signal handler and does not clean up memory. So
register a SIGTERM handler which sets a flag "worker_should_exit"
and check this flag before and after sleeps as sleeps get interrupted
by signals.
This is just to make Valgrind happy to avoid false-positive leak
reporting.