ugtrain: Use C++ string for config path
It is possible that the config path string is still located on the
stack and the cleanup tries to free it. This causes a segfault if
the path does not exist. So just convert opt->cfg_path to a C++
string, so that it is always located on the heap.

Fixes: b90ce4376323 ("ugtrain: Make Valgrind memory leak detection happy")