ugtrain: Use C++ string for process name
The process name is one of the things which has to be adapted. C++
should do the memory management so that no delete[] is missing.
Also convert the system functions to use "const char *" instead of
"char *" so that opt->proc_name->c_str() can be used when switching
to OS specific C code.