ugtrain: Use C++ string for game call
The game call is one of the things which has to be adapted. C++
should do the memory management so that no delete[] is missing.
Also convert the system functions to use "const char *" instead of
"char *" so that opt->game_call->c_str() can be used when switching
to OS specific C code.