ugtrain: Use C++ string for game binpath and add cppbasename()
The game binary path is one of the things which has to be adapted.
C++ should do the memory management so that no delete[] is missing.
Also add a new function cppbasename() which returns the file name
as C++ string from a given path. Use string::find_last_of() and
string::substr() for this. Return an empty string if the separator
cannot be found to make checks simple. Replace the basename() call
with cppbasename() in the adaptation code and make use of that
function during config parsing.