system: Optimize check_process()
Executing shell commands for this function is really lazy and ugly.
So rather use fopen() to check if /proc/$pid/status exists. If it
does not, assume that the process is dead. Then read it with
getline(), but only read the first two lines. Use offsets to extract
the required information and assume that the output format does not
change. This is good for performance but less safe. Also handle
status 'X' like 'Z' for zombie processes and check if the status is
a capital letter.

Verified with Valgrind leak detection.

Fixes #62