system: check_process: Fix for kernel >= 4.7 and leaks
With kernel >= 4.7 there is a line "Umask:	0022" between the
"Name:\t" and the "State:\t" line. That breaks the line number based
assumptions. So do string compares instead and break out of the loop
as soon as the process state is read.
Furthermore, if the process name is wrong, then this leaks the line
allocation and the file pointer. So store the state PROC_WRONG in
a temporary variable, break out of the loop and check the temporary
variable again to return PROC_WRONG when the line is freed and the
file is closed.

Fixes: 34c2cd745593 ("system: Optimize check_process()")