discovery: Read the stack end in reverse mode first
Dropping the heap start handling has the side effect that the stack
end (now in the first line) is not read any more during reverse
parsing for discovery. Handling the stack end at the end is also
always a problem - especially if many mallocs are recorded as well.

So partially revert that commit and read as well as handle the stack
end first.

Fixes: 16fefe650470 ("discovery: Drop the heap start/offset")