aslr: Fix the PIE handling
Tests on Ubuntu Bionic have shown that the commit "ugtrain: Use C++
strings for lib_name" did not cover all occurrences of dynmem->lib
and dynmem->grow->lib. This causes that offset 0x0 is sent to
libmemhack instead of the actual exe offset.
So fix the PIE handling by checking for dynmem->lib->empty() and
grow->lib->empty() as well.

Fixes: 0a8bb9b2feaf ("ugtrain: Use C++ strings for lib_name")