ugtrain: Fix a segfault without preloading
If starting the bofh64 config without -P, then this causes a segfault
as the game_path is NULL but is used to set the game_binpath. That
path does not matter if trying to attach to an already running
process.
So move that code inside the preloading check block where it is
guaranteed that the game_path is set.