util: Print error string if memattach test fails
Sometimes it is not that clear why the memory attach test fails.
So print an error string from errno in that case. Usually the issue
is "Operation not permitted" due to missing sudo when trying to
attach to an already running process on Ubuntu.
The strerror() function is platform specific. So wrap it in the
system lib functions.