completion: Add --color=never to ls command
The completion of configs does not work on openSUSE Leap 15.0 any
more when the following alias is set:
alias ls="ls --color"
In previous Leap releases, "--color=auto" was the default and now
it seems to be "--color=never" which causes that users play with
color settings.

So set "--color=never" explicitly to make the completion work
independent of the user setting.