aslr: get_stack_end: Add check for spaces in proc_name
If there are spaces in the process name, then getting the stack end
just by counting spaces will break. So search for the process name
in parentheses, count spaces in the process name, and add those
when searching the stack end. Also check if start is string::npos
to avoid an infinite loop.

Reference: #70