tools: Introduce ugstackdisc
Scanmem cannot handle forward stack offset from stack end calculation
yet. So introduce ugstackdisc for this as a POC.
The tool is useful when dynamic memory discovery cannot or is not
intended to be used for the stack offset discovery.
Parse command line options pid and value_hex_address. Then read and
parse /proc/$pid/stat. Find the exact location of the stack end by
counting spaces and handling spaces in the process name. Then
calculate the stack offset, check if it is < 16 MiB, and print it.
Also offer a debug mode for more output.

Reference: #70