discovery: init_scanmem: Fix memory leaks
When knowing already that static memory or a stack value is to be
discovered, then options "-SP" are used. The function init_scanmem()
wants to clear the config then. But it only clears the config lists.
So call clear_config() instead. Also drop the function argument as
"opt" is not needed any more. Found by Valgrind.