aslr: get_stack_end: Switch to a C implementation
Scanmem and ugstackdisc need to have the same implementation to get
the stack end from /proc/$pid/stat and to verify that the stack end
is valid. So move the function to src/lib/maps*, change functions
getting the stack region start to get the stack region end as well,
and implement the parsing based on loops reading char by char.
Introduce a STRCHR macro for this as strchr() is extremely slow.