ugstackdisc: Switch over to implementation in maps
Ugtrain switched over to a C implementation suitable for scanmem.
So make use of that implementation now, read the maps first, check
if the provided address is in range, and also move
find_stack_bounds() to maps so that the very same code can be used.

We are not interested in which regions belong to the executable here.
So keep the exe path as NULL and add support for that to the maps
handling. Only initialize and free the regions list here.

Reference: #70