examples: endless-sky: Adapt to Leap 15.1
I had to upgrade from openSUSE Leap 15.0 to 15.1. This time I used
a new manual adaptation method:

objdump -D `which endless-sky` | less

If the version is the same (0.9.6), then the object size and the
offsets within the object must be the same (3208, 0xc88). Search
for "\$0xc88" until reaching _Znwm() calls in the same area as
before. Check the next lower address, then check the next greater
return address. It is the next greater return address here which
is the very last _Znwm() call of this size. The nice part is that
the stack value remained the same.

Writing an adaptation script from this should be possible.