lib: Remove trailing semicolon in extern "C" handling
A -Wpedantic compile revealed this issue. It also found "long long",
comma at the end of enum list, issues with `container_of()` in
list.h from the Linux kernel, and the variable length array in the
memory object dump.
Those are actually needed ISO violations with advantages. E.g. the
comma at the end of the enum list allows appending a feature macro
and when appending elements, then the diff is much nicer.
So fix only the superfluous semicolon in `extern "C"` handling.