cfgparser: Add relative addressing support
In the game dangerdeep there is the torpedo vector. The start and
the end pointer must both be on the heap. For the start pointer,
a "this" check can be used. But for the end pointer, the absolute
address is required. This is hard to read and to adapt. It would
be "this+8" on 64 bit systems actually.
Also for integer arrays, like the ammo array in first-person
shooters, addressing relative to the previous config entry like
"prev+4" would come handy.

So add relative addressing based on "this+" or "this-" and "prev+"
or "prev-" followed by a decimal number or a hex number. E.g.
"this+8" and "this+0x8" are possible now. Check with isdigit() or
isxdigit() if the numbers are valid and trigger parsing errors in
case of issues. Add a macro GET_REL_ADDR() to avoid code duplication.

Fixes #68