examples: freedoom: Adapt to openSUSE Leap 15.1
While introducing the relative addressing, the freedoom2 config
has been messed up with `prev+8` for AmmoPT. It has an additional
double-barrelled shotgun. So fix this with the correct `prev+4`.

Adapt to openSUSE Leap 15.1 by rediscovering the health and the
pistol ammo values. Use the HaveXX array relative to the displayed
health, as offset 0x60 remained the same.

Fixes: 5844a8c448d6 ("examples: Make use of relative addressing")