dump: ptrmem: Print a message and dump statmem pointers
When dumping the objects in Warzone 2100, then there are no messages
that the pointer memory is dumped as well. When using the unpreferred
static memory pointers, then the pointer memory is not dumped at all.

So print a message similar to the message used for dynamic memory
and add the support for dumping memory referenced by a static
memory pointer. Introduce macros to avoid code duplication. Only
follow settled pointers != NULL.

Fixes #74
Fixes #75