ugtrain: Pass opt down to dump_all_mem_obj()
Many lists and pointers are already included in opt. It is also
required to dump the configured static memory areas.

So pass opt down in input handling to dump_all_mem_obj() and extract
all other pointers and values to reduce the number of function
parameters. Functionality remains the same.

Reference: #61