doc: statmem: Document dumping of static memory areas
With the new `dump` config entry, it is now possible to dump static
memory areas the same way like for dynamic memory and heap memory
referenced by pointers. It is possible to use the same method to
find further values nearby.
So document all that in the syntax description and in the static
memory documentation.