libmemdisc: Drop frame pointer stack limit check
In Battle for Wesnoth there is an additional stack region as a misc
region and an additional heap region as a misc region. Still _Znwm()
is used to allocate the units in the game. Discovery stage 4 does
not find the related allocations although the size is correct as the
additional stack region is more than 16 MiB from the stack end away.

So drop that stack limit check and add a debug message to show that
find_code_addresses() fails.