examples: wesnoth: Add first Battle for Wesnoth config
This game is pretty tricky. There are way too many Unit constructors.
Usually I do not add unfinished configs. But there is still quite
some work to be done for this one and I would like to show the
process better of getting to the final config.

Units are found reliably. Units not belonging to team 1 (player)
are kicked out reliably. Health, moves, and ability to attack are
modified. Experience values can only be watched for now. Also gold
modification needs more work.