Support config reference +/- number in value
In the Battle for Wesnoth it is required to set the experience to
experience max - 1 for not advancing multiple levels at once and to
keep the game stable. So allow appending +1 or -1 or any other number
to the config reference name in the value field.

Example:
ExpMax prev+12 i32 l 200 2,0 w
TEAM_CHECK
Exp prev-4 i32 l ExpMax-1 2,0 a